refactor(wow-core): improve wait strategy notification logic
- Update CommandWaitNotifier to use new notification logic
- Refactor WaitingForAfterProcessed to use isWaitingFor method
- Modify WaitingForFunction to use isWaitingForFunction extension function
- Update WaitingForProjected to use isWaitingFor method
- Add new WaitingFors file with isWaitingForFunction extension function
- Update WaitingForStage to use new waiting logic
