+++ wow-core/src/main/kotlin/me/ahoo/wow/command/wait/CommandWaitNotifier.kt
@@ -87,7 +87,7 @@ fun CommandWaitNotifier.notifyAndForget(
     waiteStrategy: EndpointWaitStrategy,
     waitSignal: WaitSignal
 ) {
-    if (!waiteStrategy.waitStrategy.shouldNotify(waitSignal.stage)) {
+    if (!waiteStrategy.waitStrategy.shouldNotify(waitSignal)) {
         return
     }
     notifyAndForget(waiteStrategy.endpoint, waitSignal)
