+++ wow-core/src/main/kotlin/me/ahoo/wow/command/wait/stage/WaitingForAfterProcessed.kt
@@ -15,10 +15,9 @@ package me.ahoo.wow.command.wait.stage
 
 import me.ahoo.wow.command.wait.CommandStage
 import me.ahoo.wow.command.wait.WaitSignal
-import me.ahoo.wow.command.wait.WaitSignalShouldNotifyPredicate
 import reactor.core.publisher.Mono
 
-abstract class WaitingForAfterProcessed : WaitingForStage(), WaitSignalShouldNotifyPredicate {
+abstract class WaitingForAfterProcessed : WaitingForStage() {
     @Volatile
     private var processedSignal: WaitSignal? = null
 
@@ -35,7 +34,7 @@ abstract class WaitingForAfterProcessed : WaitingForStage(), WaitSignalShouldNot
         super.complete()
     }
 
-    override fun shouldNotify(signal: WaitSignal): Boolean {
+    open fun isWaitingFor(signal: WaitSignal): Boolean {
         return signal.stage == stage
     }
 
@@ -54,7 +53,7 @@ abstract class WaitingForAfterProcessed : WaitingForStage(), WaitSignalShouldNot
         if (signal.stage == CommandStage.PROCESSED) {
             processedSignal = signal
         }
-        if (shouldNotify(signal)) {
+        if (isWaitingFor(signal)) {
             waitingForSignal = signal
         }
         tryComplete()
