+++ wow-core/src/main/kotlin/me/ahoo/wow/command/wait/stage/WaitingForProjected.kt
@@ -24,7 +24,7 @@ class WaitingForProjected(
     override val stage: CommandStage
         get() = CommandStage.PROJECTED
 
-    override fun shouldNotify(signal: WaitSignal): Boolean {
-        return super.shouldNotify(signal) && signal.isLastProjection
+    override fun isWaitingFor(signal: WaitSignal): Boolean {
+        return super.isWaitingFor(signal) && signal.isLastProjection
     }
 }
